순발력 게임 alert 오류 수정
