+++ src/components/speedGame/Timer.jsx
@@ -1,36 +1,15 @@
-import React, { useState, useEffect, useRef } from "react";
+import React, { useState, useEffect } from "react";
 import styled from "styled-components";
-import Swal from "sweetalert2";
 
-function Timer() {
+function Timer({ record }) {
   const [timeElapsed, setTimeElapsed] = useState(0);
-  const record = useRef();
   record.current = timeElapsed;
+
   useEffect(() => {
     const timer = setInterval(() => {
       setTimeElapsed((timeElapsed) => timeElapsed + 30);
     }, 30);
-    return () => {
-      let timerInterval;
-      Swal.fire({
-        title: "축하 드립니다 !!!",
-        html: `완료 기록 : ${record.current / 1000}초`,
-        timer: 2000,
-        timerProgressBar: true,
-        didOpen: () => {
-          Swal.showLoading();
-        },
-        willClose: () => {
-          clearInterval(timerInterval);
-        },
-      }).then((result) => {
-        if (result.dismiss === Swal.DismissReason.timer) {
-          console.log("I was closed by the timer");
-        }
-      });
-      //alert("완료 기록 :" + record.current / 1000 + "초");
-      clearInterval(timer);
-    };
+    return () => clearInterval(timer);
   }, []);
   return (
     <Container>
